# src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  title: 채용 공고 관리 API
  description: 사용자 인증, 채용 공고 관리, 지원 및 북마크 기능을 제공하는 API 문서입니다.
  version: 1.0.0
servers:
  - url: http://113.198.66.75:10217
    description: 배포된 API 서버
paths:
  /applications/{id}:
    post:
      summary: 채용 공고 지원하기
      description: 특정 채용 공고에 지원을 등록합니다.
      tags:
        - 지원 관리
      parameters:
        - name: id
          in: path
          required: true
          description: 지원할 채용 공고의 ID
          schema:
            type: integer
      responses:
        '200':
          description: 지원이 성공적으로 등록되었습니다.
        '401':
          description: 로그인이 필요합니다.
        '403':
          description: 이미 지원한 공고입니다.
        '404':
          description: 해당 채용 공고를 찾을 수 없습니다.
    delete:
      summary: 지원 취소
      description: 등록된 지원 내역을 취소합니다.
      tags:
        - 지원 관리
      parameters:
        - name: id
          in: path
          required: true
          description: 취소할 지원 내역의 ID
          schema:
            type: integer
      responses:
        '200':
          description: 지원이 성공적으로 취소되었습니다.
        '401':
          description: 로그인이 필요합니다.
        '403':
          description: 취소 권한이 없습니다.
        '404':
          description: 해당 지원 내역을 찾을 수 없습니다.
  /auth/register:
    post:
      summary: 회원가입
      description: 새로운 사용자 계정을 생성합니다.
      tags:
        - 사용자 인증
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: 사용자의 이메일 주소
                password:
                  type: string
                  description: 사용자의 비밀번호
      responses:
        '200':
          description: 회원가입이 성공적으로 완료되었습니다.
        '400':
          description: 잘못된 이메일 형식 또는 중복된 사용자 이름입니다.
  /auth/login:
    post:
      summary: 로그인
      description: 사용자가 로그인하고 JWT 토큰을 반환받습니다.
      tags:
        - 사용자 인증
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: 사용자의 이메일 주소
                password:
                  type: string
                  description: 사용자의 비밀번호
      responses:
        '200':
          description: 로그인 성공. 토큰 반환.
        '401':
          description: 잘못된 사용자 이름 또는 비밀번호입니다.
  /auth/profile:
    get:
      summary: 사용자 프로필 조회
      description: 사용자의 세부정보, 지원 내역, 관심 공고 목록을 조회합니다.
      tags:
        - 사용자 인증
      responses:
        '200':
          description: 프로필 정보를 성공적으로 조회했습니다.
        '401':
          description: 로그인이 필요합니다.
    put:
      summary: 사용자 프로필 수정
      description: 사용자 이메일 또는 비밀번호를 수정합니다.
      tags:
        - 사용자 인증
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: 새로운 이메일 주소
                password:
                  type: string
                  description: 새로운 비밀번호
      responses:
        '200':
          description: 프로필이 성공적으로 수정되었습니다.
        '400':
          description: 잘못된 이메일 형식 또는 비밀번호가 비어 있습니다.
  /jobs:
    get:
      summary: 채용 공고 목록 조회
      description: 사용 가능한 모든 채용 공고를 검색하거나 필터링하여 조회합니다.
      tags:
        - 채용 공고
      parameters:
        - name: keyword
          in: query
          required: false
          description: 검색할 키워드
          schema:
            type: string
        - name: location
          in: query
          required: false
          description: 지역별 필터링
          schema:
            type: string
        - name: pagenum
          in: query
          required: false
          description: 페이지 번호
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 채용 공고 목록 조회 성공.
        '400':
          description: 잘못된 요청 파라미터입니다.
  /jobs/{id}:
    get:
      summary: 채용 공고 상세 조회
      description: 특정 채용 공고의 세부 정보를 조회합니다.
      tags:
        - 채용 공고
      parameters:
        - name: id
          in: path
          required: true
          description: 조회할 채용 공고의 ID
          schema:
            type: integer
      responses:
        '200':
          description: 채용 공고 세부 정보 조회 성공.
        '404':
          description: 해당 채용 공고를 찾을 수 없습니다.
  /bookmarks/{id}:
    post:
      summary: 관심 공고 등록
      description: 특정 채용 공고를 관심 공고 목록에 추가합니다.
      tags:
        - 관심 공고
      parameters:
        - name: id
          in: path
          required: true
          description: 관심 등록할 공고의 ID
          schema:
            type: integer
      responses:
        '200':
          description: 관심 공고가 성공적으로 등록되었습니다.
        '401':
          description: 로그인이 필요합니다.
        '403':
          description: 이미 관심 등록된 공고입니다.
    delete:
      summary: 관심 공고 제거
      description: 관심 공고 목록에서 특정 공고를 제거합니다.
      tags:
        - 관심 공고
      parameters:
        - name: id
          in: path
          required: true
          description: 제거할 공고의 ID
          schema:
            type: integer
      responses:
        '200':
          description: 관심 공고가 성공적으로 제거되었습니다.
        '401':
          description: 로그인이 필요합니다.
        '404':
          description: 해당 관심 공고를 찾을 수 없습니다.
components:
  schemas:
    Job:
      type: object
      properties:
        id:
          type: integer
          description: 채용 공고 ID
        title:
          type: string
          description: 채용 공고 제목
        company:
          type: string
          description: 채용 기업
        location:
          type: string
          description: 채용 지역
        description:
          type: string
          description: 공고 상세 설명